pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "egd-parser-service"
version = "0.1.0"
description = "OCR microservice for parsing Moscow unified housing documents (EGD) from PDF into JSON"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.3,<3.0",
  "Pillow>=10,<12",
  "python-multipart>=0.0.9,<0.1",
  "uvicorn[standard]>=0.30,<1.0",
  "httpx>=0.27,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.2,<9.0",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.5,<1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
